pyproject.toml
[project]
name = "sturdy-journey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.0",
    "openai>=2.18.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
